Users are assigned role user by default
Using a static constant in User.php (model), 2 constants: USER and ADMIN, both holding the value of user and admin respectively. Doing this via constants to reduce the risk of typo and will always have to do "User::USER" or "User::ADMIN" so there is no way I can mistake a role.

User migration also has a default on role being "user" as a failsafe if I forget to change it or if I would change User.php by accident later.