Added avgRating to Recipe.php and made seeders
avgRating takes the reviews of a Recipe (model) and uses the built in average method and returns this, if there are no reviews this is 0 by default

Also made all the seeders that I need for now, might add more later but I think this is enough to built the website for the moment. As read in the extra features I might add a forum or other functionaliteiten.