Made login and register
Mainly left the code just how breeze generated it for me but customized the register and login (blade.php) views. To have the main (layout) and the business colors that im going for = amber as accent color because my logo is a cookie and it looks like a cookie color.

The user has the possibility to log in using their username or email which is by default only by email (with breeze)